Completed Basic Features
The code can listen to your voice with adjustments for ambient noises, take requests manually, and respond to it appropriately. With this being said, it stops after pauses and needs to have a better User Interface. Possible solutions include using pyaudio to make the voice into a .wav file and analyzing that to account for pauses.